% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{add_new_car}
\alias{add_new_car}
\title{add_new_car}
\usage{
add_new_car(automob, old_car, rep_type, carbu_type, year)
}
\arguments{
\item{automob}{An automob dataframe}

\item{old_car}{The old_car line}

\item{rep_type}{The name used in the MARQUE column}

\item{carbu_type}{The type of fuel (Ele or Ess)}

\item{year}{The year of the new car}
}
\value{

}
\description{
Adds a new vehicle to the automob dataframe based on the line of old_car
}
\examples{
add_new_car(automob, old_car, rep_type, year)
}
