% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_apply_transition_effect.R
\name{apply_ecoconduite}
\alias{apply_ecoconduite}
\title{apply_ecoconduite}
\usage{
apply_ecoconduite(effet_mat_df)
}
\arguments{
\item{effet_mat_df}{}
}
\value{
A chained effet_mat_df dataframe
}
\description{
Apply the effect of ecoconduite on the consumption of transportation fuels
}
\examples{
apply_ecoconduite(effet_mat_df)
}
