% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_apply_historical_trend_effect.R
\name{apply_elasticities}
\alias{apply_elasticities}
\title{apply_elasticities}
\usage{
apply_elasticities(
  men_elast_df,
  spending_aggr,
  price_index_hor_df,
  spending_var_sub,
  floor_at_z = T
)
}
\arguments{
\item{spending_aggr}{A spending dataframe}

\item{price_index_hor_df}{A price dataframe that provides one line with the growth factor at the horizon}

\item{spending_var_sub}{A dataframe containing 3 columns IDENT_MEN, SpendingRef the current spending at ref year, SpendingHor,
the current spending at horizon}

\item{menage_typo_df}{A dataframe containing the ident_men, typo and deciles of income}
}
\value{
Returns a transformed spending_econo
}
\description{
Applies the elasticities from menage_typo_df to the spending_econo, based on price_index and spending_var, the
tables for the price and income growth factors
}
\examples{
apply_elasticities(menage_typo_df, spending_aggr, price_index_hor_df, spending_var_sub)
}
