% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_apply_historical_trend_effect.R
\name{apply_historical_trend_effect}
\alias{apply_historical_trend_effect}
\title{apply_historical_trend_effect}
\usage{
apply_historical_trend_effect(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A modified spending_aggr dataframe
}
\description{
The step 'historical trend'
}
\examples{
apply_historical_trend_effect(MatisseData)
}
