% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_apply_transition_effect.R
\name{apply_homothetic_reduction}
\alias{apply_homothetic_reduction}
\title{apply_homothetic_reduction}
\usage{
apply_homothetic_reduction(MatisseData, effet_mat_df)
}
\arguments{
\item{effet_mat_df}{}
}
\value{
A chained effet_mat_df dataframe
}
\description{
Apply an homothetic reduction of consumption of energies
}
\examples{
apply_homothetic_reduction(MatisseData, effet_mat_df)
}
