% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{apply_housing_equipement}
\alias{apply_housing_equipement}
\title{apply_housing_equipement}
\usage{
apply_housing_equipement(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A spending_aggr tibble
}
\description{
Applies the effect of housing equipement (changes in personal house) to the MatisseData and
returns a spending tibble updated with the new spending from the equipement
}
\examples{
apply_housing_equipement(MatisseData)
}
