% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{apply_newcar_transformation}
\alias{apply_newcar_transformation}
\title{apply_newcar_transformation}
\usage{
apply_newcar_transformation(MatisseData, attribute_new_car, year)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{attribute_new_car}{}
}
\value{
A transformed MatisseData with new car inclusion
}
\description{
Applies the different transformations of VE replacement
}
\examples{
apply_newcar_transformation(MatisseData, attribute_new_car, year)
}
