% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_apply_transition_effect.R
\name{apply_teletravail_effect}
\alias{apply_teletravail_effect}
\title{@title apply_teletravail_effect
@description An internal function that applies an effect based on the transition file}
\usage{
apply_teletravail_effect(menage_sub, effet_mat_df)
}
\arguments{
\item{menage_sub}{A menage dataframe that provides the information on which the teletravail is applied}

\item{effet_mat_df}{A chained effet_mat_df dataframe}
}
\value{
A chained effet_mat_df dataframe
}
\description{
@title apply_teletravail_effect
@description An internal function that applies an effect based on the transition file
}
\examples{
apply_teletravail_effect(menage_sub, effet_mat_df)
}
