% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_apply_transition_effect.R
\name{apply_transition_effect}
\alias{apply_transition_effect}
\title{apply_transition_effect}
\usage{
apply_transition_effect(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A spending_aggr dataframe
}
\description{
Apply transition
}
\examples{
apply_transition_effect(MatisseData)
}
