% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{apply_VE_transformation}
\alias{apply_VE_transformation}
\title{apply_VE_transformation}
\usage{
apply_VE_transformation(MatisseData, attribute_VE, year)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{attribute_VE}{}
}
\value{

}
\description{
Applies the different transformations of VE replacement
}
