% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_transform_3ME_data.R
\name{calculate_3ME_conso}
\alias{calculate_3ME_conso}
\title{calculate_3ME_conso}
\usage{
calculate_3ME_conso()
}
\value{
A price dataframe with a year column and the price index for each of the Econometry sectors
}
\description{
This function returns, for the Econometry level of agregation,
the consommation per sector based on the Sorties-3ME data.
}
\examples{
calculate_3ME_conso()
}
