% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_transform_3ME_data.R
\name{calculate_3ME_price_index}
\alias{calculate_3ME_price_index}
\title{calculate_3ME_price_index}
\usage{
calculate_3ME_price_index()
}
\value{
A price dataframe with a year column and the price index for each of the Econometry sectors
}
\description{
This function returns, for the Econometry level of agregation, the price index based on the Sorties-3ME data.
It is a Fisher price index. The function will agregate the levels from 3ME into the level of Econometry based on the Transco_3ME_Econometry
file that gives agregation between the sectors of the source and the sectors of the target. The Price Index uses the price from 3ME weighted
by Domestic and Imported production.
}
\examples{
calculate_3ME_price_index()
}
