% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{calculate_cost_renov}
\alias{calculate_cost_renov}
\title{calculate_cost_renov}
\usage{
calculate_cost_renov(MatisseData, attribute_renov, year)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{attribute_renov}{The attribution tibble for renovation}

\item{year}{The year of application}
}
\value{
A renov_distri tibble
}
\description{
Calculates the cost of renovations, based on ThreeMe data, surfhab for the households that do renovate
}
\examples{
calculate_cost_renov(MatisseData, attribute_renov, year)
}
