% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{calculate_rdb}
\alias{calculate_rdb}
\title{calculate_rdb}
\usage{
calculate_rdb(menage, rev_vec)
}
\arguments{
\item{menage}{A menage dataframe}
}
\value{
A menage dataframe
}
\description{
Recalculate the RDB and RDBAI columns
}
\examples{
calculate_rdb(menage)
}
