% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{calculate_savings_rate}
\alias{calculate_savings_rate}
\title{calculate_savings_rate}
\usage{
calculate_savings_rate(savings_df, col_saving = "Savings")
}
\arguments{
\item{savings_df}{A standard savings_df}

\item{col_saving}{A col_name}
}
\value{
A saving_rate dataframe which contains the household id, RDB (Income - Taxes), the Savings column and the SavingRate
}
\description{
This function calculate the saving rate dataframe based on the savings_df entered and the col_saving which indicates
which column is used as the relevant Saving data (typically either Savings or SavingsExDurable)
}
\examples{
calculate_savings_rate(savings_df, "Savings")
}
