% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_get_budget_part.R
\name{correct_gazelec}
\alias{correct_gazelec}
\title{correct_gazelec}
\usage{
correct_gazelec(spending_aggr, MatisseData)
}
\arguments{
\item{spending_aggr}{A spending_aggr tibble}

\item{MatisseData}{A MatisseData tibble}
}
\value{

}
\description{
This function corrects the category M04.5.0 which is undissociated gaz and elec for households
that have a common bill. It is based on a simple regression based on the main source of heating.
}
\examples{
correct_gazelec(spending_aggr, MatisseData)
}
