% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{distrib_overbuild}
\alias{distrib_overbuild}
\title{distrib_overbuild}
\usage{
distrib_overbuild(constr_target)
}
\arguments{
\item{constr_target}{}
}
\value{
A tibble for the construction targets
}
\description{
An internal function to update the constr_target tibble when there's over construction
}
\examples{
distrib_overbuild(constr_target)
}
