% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_utils.R
\name{easy_stats}
\alias{easy_stats}
\title{easy_stats}
\usage{
easy_stats(x, stats_to_show = 1:2)
}
\arguments{
\item{x}{A vector}

\item{stats_to_show}{The type of data we want}
}
\value{

}
\description{
Returns a list of standard statistics about the vec input
}
\examples{
easy_stats(1:10)
}
