% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{extract_dummies}
\alias{extract_dummies}
\title{extract_dummies}
\usage{
extract_dummies(.data = data.frame(), col_name = "", recode_chr = "")
}
\arguments{
\item{.data}{A dataframe from which we want a dummies df for column col_name, with recode_chr being the
recode according to car::recode}

\item{col_name}{The column from which we want to extract the dummies}

\item{recode_chr}{The recode character string used for car::recode}
}
\value{
A dataframe of indicatrix
}
\description{

}
\examples{
extract_dummies(menage, "ZEAT", "1:2 = 1 ; 3 = 3 ; 4 = 4 ; 5 = 5 ; 6 = 6 ; 7 = 7 ; 8 = 8 ; 9 = 9")
}
