% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_utils.R
\name{floor_by_value}
\alias{floor_by_value}
\title{floor_by_value}
\usage{
floor_by_value(vector, value = 0)
}
\arguments{
\item{vector}{A numeric vector}

\item{value}{The floor value (defaults at 0)}
}
\value{
A floored vector
}
\description{
This function takes a numeric vector and floors the individual value by the parameter value
or by default 0
}
\examples{
a<- -5:10
floor_by_value(a)
}
