% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_generate_compare_pdf.R
\name{generate_compare_pdf}
\alias{generate_compare_pdf}
\title{generate_compare_pdf}
\usage{
generate_compare_pdf(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{

}
\description{
Produces the report in pdf format
}
\examples{
generate_compare_pdf(MatisseData)
}
