% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_auto_conso_proj}
\alias{get_auto_conso_proj}
\title{get_auto_conso_proj}
\usage{
get_auto_conso_proj()
}
\value{
A auto_conso_proj tibble
}
\description{
Return a tibble that contains the consumption for personnal vehicles, based on type of motor
expressed in Euros-horizon
}
\examples{
get_auto_conso_proj()
}
