% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_auto_proj}
\alias{get_auto_proj}
\title{get_auto_proj}
\usage{
get_auto_proj()
}
\value{
A Auto_proj dataframe
}
\description{
A function that extracts the data for the parc of cars from ThreeME
}
\examples{
get_auto_proj()
}
