% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{get_car_to_replace}
\alias{get_car_to_replace}
\title{get_car_to_replace}
\usage{
get_car_to_replace(automob, FuelType, IdMen)
}
\arguments{
\item{automob}{An automob dataframe (list of vehicles)}

\item{FuelType}{Type of fuel (can be a vector)}

\item{IdMen}{ID of the houshold}
}
\value{
An index of the vehicle to be changed
}
\description{
Returns the idx in automob for the car to replace, based on the type of fuel and the ID of the household
}
\examples{
get_car_to_replace(automob_sub, "Ele", IdMen = "00001")
}
