% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{get_constr_target}
\alias{get_constr_target}
\title{get_constr_target}
\usage{
get_constr_target(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A constr_target tibble with the DPE_from DPE_to and the surface to be constructed
}
\description{
Internal function that returns the split of constructions based on the DPE from
which the household starts and the DPE at which it arrives. Favors lesser DPE jump.
Not to be used outside of code
}
\examples{
get_constr(MatisseData)
}
