% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_csv_data}
\alias{get_csv_data}
\title{get_csv_data}
\usage{
get_csv_data(to_include = NULL)
}
\arguments{
\item{to_include}{A vector of the type of file to extract
Valid values : c("pop_M_csv", "pop_F_csv", "men_proj", "transco_sect", "transco_rev",
"typo_men", "sorties3me", "transco_3me_econo", "transition", "transco_sect_econo")
Empty value returns all valid values}
}
\value{
A list of tibbles extracting the data from specified csv files.
}
\description{
Extracts data from all MatisseFiles$bdf_***_csv files, in csv format
}
\examples{
get_csv_data()
get_csv_data(to_include = "men_proj")
}
