% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_transform_final_data.R
\name{get_dec_hor}
\alias{get_dec_hor}
\title{get_dec_hor}
\usage{
get_dec_hor(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{

}
\description{
Produces the new Déciles of revenus (based on RDB)
}
\examples{
get_dec_hor(MatisseData)
}
