% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_prepare_data.R
\name{get_dom_conso_proj}
\alias{get_dom_conso_proj}
\title{get_dom_conso_proj}
\usage{
get_dom_conso_proj(MatisseData, years = c())
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
a dom_conso tibble
}
\description{
Gets the average domestic consumption per year, per class and energy according to ThreeMe data
}
\examples{
get_dom_conso_proj(MatisseData, 2017)
}
