% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_ener}
\alias{get_ener}
\title{get_ener}
\usage{
get_ener(MatisseData, year)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{year}{The year at which we are calculating the consumption}
}
\value{
An ener tibble
}
\description{
Return an ener tibble with the domestic energy consumption for the 5 energy sources for house
Gaz, Elec, Fioul, Solide, AutreEner. Energy is counted in current E, E/m², kWh and kWh/m²
}
\examples{
get_ener(MatisseData, 2017)
}
