% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_ener_3me}
\alias{get_ener_3me}
\title{get_ener_3me}
\usage{
get_ener_3me(years = c())
}
\arguments{
\item{years}{The years on which to extract the data}
}
\value{
A transformed tibble for the domestic energy consumption from ThreeMe
}
\description{
Gets the domener data from Threeme : per energy, per year, volumes in kWh, price in ME/kWh,
total in current M€. Autres is the sum of Solides and Petrole weighted per volume
}
\examples{
get_ener_3me(2017)
}
