% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_transform_3ME_data.R
\name{get_ener_var_3Me}
\alias{get_ener_var_3Me}
\title{get_ener_var_3Me}
\usage{
get_ener_var_3Me()
}
\value{
A tibble containing ThreeMe's variation in energy for Elec, Gaz and Fuel
}
\description{
Loads the table for energy consumption from 3ME
}
\examples{
get_ener_var_3Me()
}
