% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_ener_var_Matisse}
\alias{get_ener_var_Matisse}
\title{get_ener_var_Matisse}
\usage{
get_ener_var_Matisse(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{
A Matisse ener var tibble
}
\description{
Returns the energy consumption variation for the current MatisseData
}
\examples{
get_ener_var_Matisse(MatisseData)
}
