% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_house_proj}
\alias{get_house_proj}
\title{get_house_proj}
\usage{
get_house_proj()
}
\value{
A House_proj dataframe
}
\description{
A function that extracts the data for the stock of houses from ThreeME
}
\examples{
get_house_proj()
}
