% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_intermed_spend_aggr}
\alias{get_intermed_spend_aggr}
\title{get_intermed_spend_aggr}
\usage{
get_intermed_spend_aggr(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list}
}
\value{
A list of spending data
}
\description{
Returns the intermed spending_aggr stocked in MatisseData
}
\examples{
get_intermed_spend_aggr(MatisseData)
}
