% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_IPStone}
\alias{get_IPStone}
\title{get_IPStone}
\usage{
get_IPStone(spending_aggr, price_index_hor_df)
}
\arguments{
\item{spending_aggr}{A spending dataframe at the level of Econometry}

\item{price_index_hor_df}{A price dataframe with only the year}
}
\value{
Returns an IPStone vector per household
}
\description{
Gives the Stone price index for each household based on the spending_aggr which contains the numeric spending per sector
and the price_index_hor_df which is an extract of the price index dataframe for the target year
}
\examples{
get_IPStone(spending_aggr, price_index_hor_df)
}
