% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_men_elast_df}
\alias{get_men_elast_df}
\title{get_men_elast_df}
\usage{
get_men_elast_df(menage)
}
\arguments{
\item{menage}{A menage dataframe}
}
\value{
A men_elast dataframe
}
\description{
Create a men_elast dataframe which contains for each household the values for the elasticities, the decile, the
typo
}
\examples{
get_men_elast_df(menage)
}
