% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_men_proj}
\alias{get_men_proj}
\title{get_men_proj}
\usage{
get_men_proj(years)
}
\arguments{
\item{years}{The projected year for which we want the buckets}
}
\value{
Return a tibble of buckets by year of projection, age of bucket (from and to, including both), the sex and sum value
}
\description{
Extracts from the menage projection file the number of households,
including by typmen5 (3 and 4 merged) for the years#'
}
\examples{
get_men_proj(years = c(2017,2030))
}
