% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_newauto_th_conso_proj}
\alias{get_newauto_th_conso_proj}
\title{get_newauto_th_conso_proj}
\usage{
get_newauto_th_conso_proj()
}
\value{
A newauto_conso_proj tibble
}
\description{
Return a tibble that contains the average consumption for new thermic vehicles, based on type of motor
expressed in Euros-2006
}
\examples{
get_newauto_th_conso_proj()
}
