% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_parc_auto}
\alias{get_parc_auto}
\title{get_parc_auto}
\usage{
get_parc_auto(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{

}
\description{
A function that recalculates a few columns in vehic (NbVehic_XXX and min/max years of cars)
}
\examples{
get_parc_auto(MatisseData)
}
