% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_pop_proj}
\alias{get_pop_proj}
\title{get_pop_proj}
\usage{
get_pop_proj(years, buckets = NULL)
}
\arguments{
\item{years}{The projected year for which we want the buckets}

\item{buckets}{The limits for the bucketting of projected population}
}
\value{
Return a tibble of buckets by year of projection, age of bucket (from and to, including both), the sex and sum value
}
\description{
Extracts from the population projection file the number of individuals for the years
in input, split by sex and by buckets
}
\examples{
get_pop_proj(years = c(2017,2030) ,  buckets = c(15,30,45,60,75))
}
