% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{get_ranking_build}
\alias{get_ranking_build}
\title{get_ranking_build}
\usage{
get_ranking_build(MatisseData, type = "Cost", dec_inc = "dec")
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{type}{A string for the ranking metric. Defaults at 'Cost'
Accepted values are:
'Cost' uses the monetary spending all domestic energies
'Ener' uses the kWh consumption in all domestic energies
'ConsoSurf' uses the kWh/m2 consumption in all domestic energies
'Fossile'uses the sum of kWh consumption in fossile energies}

\item{dec_inc}{A string (valid dec or inc) indicating if the type should be ordered increasingly or decreasingly.
In the natural process, low ranking will be favored for selection. Thus a 'Cost dec' type + dec_inc will
result in the higher monetary spending be favored."Random" uses the default ranking of household (the line number of each)
to generate a pseudorandom but deterministic allocation of tech change.}
}
\value{
A ranking tibble
}
\description{
A function to rank the household for the renovation.
The ranking is done per class of household, and favors based on the status of the house :
In order of priority :
}
\examples{
get_ranking_build(MatisseData, type = "Cost", dec_inc = "dec")
}
