% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{get_ranking_transport}
\alias{get_ranking_transport}
\title{get_ranking_transport}
\usage{
get_ranking_transport(MatisseData, type = "Cost", dec_inc = "dec")
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{type}{A string for the ranking metric. Defaults at 'Cost'
Accepted values are:
'Cost' uses the monetary spending in transport fuel}

\item{dec_inc}{A string (valid dec or inc) indicating if the type should be ordered increasingly or decreasingly.
In the natural process, low ranking will be favored for selection. Thus a 'Cost dec' type + dec_inc will
result in the higher monetary spending be favored.}
}
\value{

}
\description{

}
\examples{
get_ranking_transport(MatisseData, type = "Cost", dec_inc = "dec")
}
