% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{get_ratio_m2_hor}
\alias{get_ratio_m2_hor}
\title{get_ratio_m2_hor}
\usage{
get_ratio_m2_hor(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{
A ratio of m2 at horizon to apply to expected m2, in order to align ThreeMe targets with Matisse
}
\description{
Internal function returning the ratio of m2 at horizon between ThreeMe and Matisse
}
\examples{
get_ratio_m2_hor(MatisseData)
}
