% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_renov_cost_proj}
\alias{get_renov_cost_proj}
\title{get_renov_cost_proj}
\usage{
get_renov_cost_proj()
}
\value{
A renov_copst dataframe
}
\description{
A function that extracts the data for the renovation costs from ThreeME
}
\examples{
get_renov_cost_proj()
}
