% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_sas_data}
\alias{get_sas_data}
\title{get_sas_data}
\usage{
get_sas_data(to_include = NULL)
}
\arguments{
\item{to_include}{A vector of the data we want to extract.
Empty value returns every data available.
Valid values : c("menage", "c05", "indiv", "depindiv", "depmen", "automob")}
}
\value{
A list of dataframes extracting the data from BDF sas files. Names are the 'to_include' valid name
}
\description{
Extracts the bdf data from all MatisseFiles$bdf_***_sas files, in sas7 format.
}
\examples{
list_res <- get_sas_data()
list_res <- get_sas_data(to_include = "menage")
}
