% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_get_budget_part.R
\name{get_savings}
\alias{get_savings}
\title{get_savings}
\usage{
get_savings(MatisseData, type = "ref")
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A savings dataframe
}
\description{
This function extracts a standard savings dataframe from the income, taxes, spending and durable dataframe
}
\examples{
get_savings(MatisseData)
}
