% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_savings_proj}
\alias{get_savings_proj}
\title{get_savings_proj}
\usage{
get_savings_proj(years)
}
\arguments{
\item{years}{}
}
\value{
Return a tibble of the saving rate by year of projection
}
\description{
Extracts the projected saving rate from ThreeMe
}
\examples{
get_savings_proj(c(2017,2035)
}
