% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_taxe_proj}
\alias{get_taxe_proj}
\title{get_taxe_proj}
\usage{
get_taxe_proj(years)
}
\arguments{
\item{years}{The projected year for which we want the buckets}
}
\value{
Return a tibble of buckets by year of projection, and type of taxes
}
\description{
Returns a tibble with the projected taxes
}
\examples{
get_taxe_proj(c(2017,2035))
}
