% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_get_budget_part.R
\name{get_taxes}
\alias{get_taxes}
\title{get_taxes}
\usage{
get_taxes(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A standard taxes dataframe
}
\description{
This function extracts the taxes data from the menage and c05 dataframe to the standard taxes dataframe.
}
\examples{
get_taxes(MatisseData)
}
