% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_teletravail}
\alias{get_teletravail}
\title{get_teletravail}
\usage{
get_teletravail()
}
\value{
A tibble extract of the transition file
}
\description{
Returns the data from the transition file
}
\examples{
get_teletravail()
}
