% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_threeme_data}
\alias{get_threeme_data}
\title{get_threeme_data}
\usage{
get_threeme_data(years = NULL, fields = NULL)
}
\arguments{
\item{years}{A vector of years to return data from}

\item{fields}{A vector of the fields to return}
}
\value{
A tibble with 3 columns : variable, year and value of the variable on given year
}
\description{
Extracts from the sorties_3me file the data into a 3 col dataframe : variable, year, value.
Defaults extract all years and variables, specifying fields and years (vector entry) to only return the specified
values
}
\details{
get_threeme_data
}
\examples{
get_threeme_data()
get_threeme_data(years = c("2015",2030))
get_threeme_data(years = c("2015",2030), fields = c("POP_TOT", "Invalid", "VA_06_2"))
}
