% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_vehic}
\alias{get_vehic}
\title{get_vehic}
\usage{
get_vehic(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{
A vehic tibble which contains informations about the vehicules and transportation for all households
}
\description{
Extract the vehic dataframe which summarizes the ownership of cars by the household
}
\examples{
get_vehic(MatisseData)
}
