% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{get_VT_efficiency}
\alias{get_VT_efficiency}
\title{get_VT_efficiency}
\usage{
get_VT_efficiency(MatisseData)
}
\arguments{
\item{years}{}
}
\value{

}
\description{
Returns a tibble with the evolution of average consumption for VT, which is then used to adjust
fuel consumption for VT on change of equipment
}
\examples{
get_VT_efficiency(MatisseParams$year_ref:MatisseParams$year_hor)
}
