% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_read_files.R
\name{get_work_proj}
\alias{get_work_proj}
\title{get_work_proj}
\usage{
get_work_proj(years)
}
\arguments{
\item{years}{The years to target}
}
\value{
A work_proj tibble
}
\description{
Returns the rate of active and unemployed to population
}
\examples{
get_work_proj(years)
}
