% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_initialize_Matisse.R
\name{initialize_MatisseParams}
\alias{initialize_MatisseParams}
\title{initialize_MatisseParams}
\usage{
initialize_MatisseParams(
  pop_age_buckets = c(15, 30, 45, 60, 75),
  year_ref = 2017,
  year_hor = 2035,
  scenario = "S1",
  classement_dom = "Cost_dec",
  classement_veh = "Cost_dec",
  save_intermed_file = FALSE
)
}
\arguments{
\item{pop_age_buckets}{The buckets of age pop split (for repond)}

\item{year_ref}{The year considered as the reference}

\item{year_hor}{The year for the end of the simulation. Standard values : 2035, 2050}

\item{scenario}{The scenario from 3ME : values S1->S4}

\item{classement_dom}{The way to select households to have a technical change for houses (Cost_dec standard)}

\item{classement_veh}{The way to select households to switch to electric vehicule (Cost_dec standard)}

\item{save_intermed_file}{Should intermediary files be saved}
}
\value{
TRUE but creates MatisseParams, a global variable, a list that contains all the params
}
\description{
The initial setting of MatisseParams
}
\examples{
initialize_MatisseParams(scenario = "S1", year_hor = 2035)
initialize_MatisseParams()
}
