% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{is_eligible_constr}
\alias{is_eligible_constr}
\title{is_eligible_constr}
\usage{
is_eligible_constr(MatisseData, is_horizon = FALSE)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{is_horizon}{A boolean indicating whether the check is for horizon year or not}
}
\value{
A elig_constr tibble
}
\description{
A function that applies the rules to say if a given household can have
a new construction. Current conditions are...
}
\examples{
is_eligible_constr(MatisseData, is_horizon = T)
}
