% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{is_eligible_new_VE}
\alias{is_eligible_new_VE}
\title{is_eligible_new_VE}
\usage{
is_eligible_new_VE(MatisseData, year)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{year}{The year of appication of the filter}
}
\value{
A elig_new_VE tibble
}
\description{
A function
}
\examples{
is_eligible_new_VE(MatisseData, year_it)
}
