% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_transport_equipement.R
\name{is_eligible_new_VT}
\alias{is_eligible_new_VT}
\title{is_eligible_new_VT}
\usage{
is_eligible_new_VT(MatisseData, year)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{year}{The year of appication of the filter}
}
\value{
A elig_new_VT tibble
}
\description{
A function that applies the rules to say if a given household can buy a new VT
}
\examples{
is_eligible_new_VT(MatisseData, year_it)
}
