% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Matisse_Launcher}
\alias{Matisse_Launcher}
\title{Matisse_Launcher}
\usage{
Matisse_Launcher()
}
\arguments{
\item{MatisseData}{}
}
\value{
A MatisseData list
}
\description{
The main launcher of Matisse
}
\examples{
Matisse_Launcher()
}
