% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_utils.R
\name{max_na}
\alias{max_na}
\title{max_na}
\usage{
max_na(x = c(), ...)
}
\arguments{
\item{x}{A vector}

\item{...}{A free arg}
}
\value{
A vector or NA
}
\description{
A finer max that handles situations where the vector is empty or full of NA
}
\examples{
max_na(c())
max_na(c(NA, 10), na.rm = T)
}
