% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_utils.R
\name{min_na}
\alias{min_na}
\title{min_na}
\usage{
min_na(x = c(), ...)
}
\arguments{
\item{x}{A vector}

\item{...}{A free arg}
}
\value{
A vector or NA
}
\description{
A finer min that handles situations where the vector is empty or full of NA
}
\examples{
min_na(c())
min_na(c(NA, 10), na.rm = T)
}
