% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{modify_budget_housing_equipement}
\alias{modify_budget_housing_equipement}
\title{modify_budget_housing_equipement}
\usage{
modify_budget_housing_equipement(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A spending_aggr tibble
}
\description{
A function that modifies the budgets for households based on the equipement changes, including the
ventilation of budget residues
}
\examples{
modify_budget_housing_equipement(MatisseData)
}
