% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{na_to_na}
\alias{na_to_na}
\title{na_to_na}
\usage{
na_to_na(x, NA_codes = c())
}
\arguments{
\item{x}{A single vector X}

\item{NA_codes}{A vector of NA_codes. If left as default, uses the most common list of NA values instead}
}
\value{
Return a vector of the same size as x with NA instead of other NA values
}
\description{
A very simple function that replaces the most common NA values from different sources (Excel, csv,...) into
R NA. Internal only, to use with caution.
}
\examples{
na_to_na(c("NA", "N/A", "Unchanged", "Matisse", 1))
}
