% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_prepare_data.R
\name{prepare_automob}
\alias{prepare_automob}
\title{prepare_automob}
\usage{
prepare_automob(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{
An automob tibble
}
\description{
Internal function to produce a simpler automob tibble
}
\examples{
prepare_automob(MatisseData)
}
