% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_prepare_data.R
\name{prepare_DPE}
\alias{prepare_DPE}
\title{prepare_DPE}
\usage{
prepare_DPE(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A DPE tibble
}
\description{
Produces the initial DPE classification based on domestic energy consumption
}
\examples{
prepare_DPE(MatisseData)
}
