% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_prepare_data.R
\name{prepare_menage}
\alias{prepare_menage}
\title{prepare_menage}
\usage{
prepare_menage(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}
}
\value{
A menage dataframe
}
\description{
This function cleans up the data, removing non-handled households. In particular, it
\itemize{
\item removes ZEAT = 0, i.e. DOM regions
}
}
\examples{
prepare_menage(MatisseData)
}
