% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_adjust_budget.R
\name{project_incomes}
\alias{project_incomes}
\title{project_incomes}
\usage{
project_incomes(MatisseData, inc_proj)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{inc_proj}{A dataframe containing the value (current value) for the different incomes categories}
}
\value{
An income dataframe that contains the projected income
}
\description{
Step 2 : adjust the income  for the households budgets
}
\examples{
project_incomes(MatisseData, inc_proj)
}
