% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_adjust_budget.R
\name{project_taxes}
\alias{project_taxes}
\title{project_taxes}
\usage{
project_taxes(MatisseData, tax_proj)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{tax_proj}{A dataframe containing the value (current value) for the different taxes categories}
}
\value{
A taxes dataframe
}
\description{
This function updates the taxes per household based on the tax_proj dataframe which provides the level for
the ref year and the year_hor. It adjusts by the growth of the population based on pondmen.
}
\examples{
project_taxes(MatisseData, tax_proj)
}
