% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_utils.R
\name{replace_nan}
\alias{replace_nan}
\title{replace_nan}
\usage{
replace_nan(x = c(), by = NA, ...)
}
\arguments{
\item{x}{An input object}

\item{by}{The replacement for NaN values}

\item{...}{}
}
\value{
An object
}
\description{
Function that replaces NaN values with "by" parameter
}
\examples{
replace_nan(c(NA, NaN, 1,2,3), by = 0)
}
