% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_reweight.R
\name{reweight_population}
\alias{reweight_population}
\title{reweight_population}
\usage{
reweight_population(
  MatisseData,
  pop_proj,
  men_proj,
  surf_proj,
  auto_proj,
  constraints = NULL
)
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{pop_proj}{The projected population (columns : age_fr, age_to, sex, year, value)}

\item{men_proj}{The projected households (columns : type, year, value)}

\item{auto_proj}{The projected vehicle (columns : year, value)}

\item{constraints}{The list of columns on which the calibration is done. Defaults include all valid values.
Valid values : c("TYPMEN5", "ZEAT", "AGEPR", "TUU", "VAG", "AgeSex", "SurfHab", "NbVehic")}
}
\value{
Return the a pondmen_df datafram containing just the old and new ponderations
}
\description{
This function does a reweight of the population from the input menage_sub dataframe,
based on the indiv_sub dataframe, the pop_proj and men_proj values for estimated modification of the population
and households.
}
\examples{
reweight_population(MatisseData, pop_proj, men_proj, constraints = c("TYPMEN5", "TUU"))

}
