% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{split_rank_opt}
\alias{split_rank_opt}
\title{split_rank_opt
@description Splits the rank argument into a list}
\usage{
split_rank_opt(x, sep = "_")
}
\arguments{
\item{x}{A string}
}
\value{
A list
}
\description{
split_rank_opt
@description Splits the rank argument into a list
}
\examples{
split_rank_opt("Cost_dec")
}
