% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_0_Extract_All_Data}
\alias{Step_0_Extract_All_Data}
\title{Step_0_Extract_All_Data}
\usage{
Step_0_Extract_All_Data()
}
\value{
A MatisseData list
}
\description{
Extract all data needed for Matisse.
}
\examples{
Step_0_Extract_All_Data()
}
