% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_1_Reweight}
\alias{Step_1_Reweight}
\title{Step_1_Reweight}
\usage{
Step_1_Reweight(MatisseData)
}
\value{
A MatisseData list
}
\description{
Reweight the data from Matisse.
}
\examples{
Step_1_Reweight(MatisseData)
}
