% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_2_ProjectMacro}
\alias{Step_2_ProjectMacro}
\title{Step_2_ProjectMacro}
\usage{
Step_2_ProjectMacro(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list}
}
\value{

}
\description{
Project the macro elements on households
}
\examples{
Step_2_ProjectMacro(MatisseData)
}
