% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_3_HistoricalTrends}
\alias{Step_3_HistoricalTrends}
\title{Step_3_HistoricalTrends}
\usage{
Step_3_HistoricalTrends(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list}
}
\value{
A MatisseData list
}
\description{
Applies the historical trends effects (changes in income, prices affect spending)
}
\examples{
Step_3_HistoricalTrends(MatisseData)
}
