% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_4_TransitionEffects}
\alias{Step_4_TransitionEffects}
\title{Step_4_TransitionEffects}
\usage{
Step_4_TransitionEffects(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list}
}
\value{
A MatisseData list
}
\description{
Applies the transition effects (changes in behaviours)
}
\examples{
Step_4_TransitionEffects(MatisseData)
}
