% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_5_EquipementEffects}
\alias{Step_5_EquipementEffects}
\title{Step_5_EquipementEffects}
\usage{
Step_5_EquipementEffects(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list}
}
\value{
A MatisseData list
}
\description{
Applies the equipement effects (changes in house and transportation equipement)
}
\examples{
Step_5_EquipementEffects(MatisseData)
}
