% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_agecat}
\alias{summarise_agecat}
\title{summarise_agecat}
\usage{
summarise_agecat(MatisseData, scenario, split_by = c())
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vector}
}
\value{
A tibble with average age
}
\description{
A summary for age category
}
\examples{
summarise_agecat(MatisseData, "Ref")
}
