% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_auto_count}
\alias{summarise_auto_count}
\title{summarise_auto_replace}
\usage{
summarise_auto_count(MatisseData, scenario, split_by = c(), subset_idx = c())
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vector}

\item{subset_idx}{A subset of MatisseData}
}
\value{
A Car count tibble
}
\description{
Summarises the number of cars
}
\examples{
summarise_auto_count(MatisseData, scenario, split_by = c(), subset_idx = c())
}
