% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_demography}
\alias{summarise_demography}
\title{summarise_demography}
\usage{
summarise_demography(MatisseData, scenario, split_by = c())
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vector}
}
\value{
A tibble with average age
}
\description{
A summary for demographics data
}
\examples{
summarise_demography(MatisseData, "Ref")
}
