% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_enerdep}
\alias{summarise_enerdep}
\title{summarise_enerdep}
\usage{
summarise_enerdep(
  MatisseData,
  scenario,
  split_by = c(),
  subset_idx = c(),
  price_deflator = 1
)
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{Scenario}

\item{split_by}{A column of menage to split}

\item{subset_idx}{An index of lines to be included}

\item{price_deflator}{A price deflator}
}
\value{
A tibble with the sum of spending in euros
}
\description{
Summarise energy spending
}
\examples{
summarise_enerdep(MatisseData, scenario, split_by = c(), subset_idx = c(),price_deflator = 1)
}
