% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_enerphys}
\alias{summarise_enerphys}
\title{summarise_enerphys}
\usage{
summarise_enerphys(
  MatisseData,
  price_vol_df,
  scenario,
  split_by = c(),
  subset_idx = c()
)
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{price_vol_df}{A price vol tibble}

\item{scenario}{A scenario}

\item{split_by}{A column of menage to split}

\item{subset_idx}{An index of lines to be included}
}
\value{
A tibble with the sum of spendings
}
\description{
Summarises the physical energy spending
}
\examples{
summarise_enerphys(MatisseData, price_vol_df, scenario, split_by = c(), subset_idx = c(), price_deflator = 1)
}
