% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_enerphys_gini}
\alias{summarise_enerphys_gini}
\title{summarise_enerphys_gini}
\usage{
summarise_enerphys_gini(MatisseData, price_vol_df, scenario, ener_vec)
}
\arguments{
\item{MatisseData}{A MatisseData}

\item{price_vol_df}{A price_vol tibble}

\item{scenario}{A scenario}
}
\value{
A Gini tibble
}
\description{
Calculate the Gini for energy consumption
}
\examples{
summarise_enerphys_gini(MatisseData, price_vol_df, scenario, ener_vec)
}
