% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_gini_curve}
\alias{summarise_gini_curve}
\title{summarise_gini_curve}
\usage{
summarise_gini_curve(weight, value, nbpoints)
}
\arguments{
\item{weight}{The weight vector}

\item{value}{The value vector}

\item{nbpoints}{Param nbpoint}
}
\value{
A tibble of cumsum pond and value
}
\description{
summarise_gini_curve
}
\examples{
summarise_gini_curve(weight, value, nbpoints)
}
