% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_house_renov}
\alias{summarise_house_renov}
\title{summarise_house_renov}
\usage{
summarise_house_renov(MatisseData, scenario, split_by = c(), subset_idx = c())
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vector}

\item{subset_idx}{A subset of MatisseData}
}
\value{
A SurfHab tibble
}
\description{
Summarises the housing renovation
}
\examples{
summarise_house_renov(MatisseData, scenario, split_by = c(), subset_idx = c())
}
