% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_income}
\alias{summarise_income}
\title{summarise_income}
\usage{
summarise_income(
  MatisseData,
  scenario,
  split_by = c(),
  subset_idx = c(),
  price_deflator = 1,
  pondref = F
)
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario value}

\item{split_by}{A column of menage to split}

\item{subset_idx}{An index of lines to be included}

\item{price_deflator}{A price deflator}
}
\value{
A tibble with the sum of incomes
}
\description{
Summarises the income values
}
\examples{
summarise_income(MatisseData, scenario, split_by = c(), subset_idx = c(), price_deflator = 1)
}
