% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_sensi_ener}
\alias{summarise_sensi_ener}
\title{summarise_sensi_ener}
\usage{
summarise_sensi_ener(
  MatisseData,
  scenario,
  split_by = c(),
  subset_idx = c(),
  level_sens = 0.1
)
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vec}

\item{subset_idx}{A subset vec}
}
\value{
A tibble of the sensierability
}
\description{
Calculates the energetic sensierability (more than 10\% RDB for dom, 10\% RDB for transport)
}
\examples{
summarise_sensi_ener(MatisseData, scenario)
}
