% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_spending}
\alias{summarise_spending}
\title{summarise_spending}
\usage{
summarise_spending(
  MatisseData,
  scenario,
  split_by = c(),
  subset_idx = c(),
  price_deflator = 1
)
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario value}

\item{split_by}{A column of menage to split}

\item{subset_idx}{An index of lines to be included}

\item{price_deflator}{The value of the deflator to ref year}
}
\value{
A tibble with the sum of spending
}
\description{
A function to return the values used for graphs
}
\examples{
summarise_spending(MatisseData, scenario, split_by = c(), subset_idx = c(), price_deflator = 1)
}
