% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_taxes_split}
\alias{summarise_taxes_split}
\title{summarise_taxes_split}
\usage{
summarise_taxes_split(
  MatisseData,
  scenario,
  split_by = c(),
  subset_idx = c(),
  price_deflator = 1
)
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vec}

\item{subset_idx}{A subset vec}

\item{price_deflator}{A price deflator value}
}
\value{
A Taxes Split tibble
}
\description{
Summarise between revenue taxes and other direct taxes
}
\examples{
summarise_taxes_split(MatisseData, "Ref")
}
