% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summarise_unemployment}
\alias{summarise_unemployment}
\title{summarise_unemployment}
\usage{
summarise_unemployment(MatisseData, scenario, split_by = c())
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vec}

\item{subset_idx}{A subset vec}
}
\value{
A tibble of the unemployment rate
}
\description{
Calculates the unemployment rate
}
\examples{
summarise_unemployment(MatisseData, scenario)
}
