% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_summary.R
\name{summary_mean_elast}
\alias{summary_mean_elast}
\title{summary_mean_elast}
\usage{
summary_mean_elast(MatisseData, scenario, split_by = c())
}
\arguments{
\item{MatisseData}{A MatisseData list}

\item{scenario}{A scenario}

\item{split_by}{A split vector}
}
\value{
A tibble with average elasticities
}
\description{
A summary for mean elastcities
}
\examples{
summary_mean_elast(MatisseData, "Ref")
}
