% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_transform_final_data.R
\name{transform_final_data}
\alias{transform_final_data}
\title{transform_final_data}
\usage{
transform_final_data(MatisseData)
}
\arguments{
\item{MatisseData}{}
}
\value{
A MatisseData
}
\description{
Adds the final data needed
}
\examples{
transform_final_data(MatisseData)
}
