% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_apply_housing_equipement.R
\name{update_DPE}
\alias{update_DPE}
\title{update_DPE}
\usage{
update_DPE(MatisseData, year, type = "renov")
}
\arguments{
\item{MatisseData}{A MatisseData list of all the data from Matisse}

\item{year}{The year of renovation}

\item{type}{The type of construction change (values = "renov" or "constr")}
}
\value{
A DPE tibble that contains all the domestic energy consumptions
}
\description{
This function updates the DPE tibble from MatisseData based on the renov_distri tibble in MatisseData,
DPE and the efficiency gains from ThreeMe
}
\examples{
update_DPE(MatisseData, year)
}
