% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{ventilate_solde}
\alias{ventilate_solde}
\title{ventilate_solde}
\usage{
ventilate_solde(
  spending_econo_df,
  men_elast_df,
  spending_var_df,
  price_index_hor_df,
  floor_at_z = T
)
}
\arguments{
\item{spending_econo_df}{A spending dataframe}

\item{men_elast_df}{A men_elast dataframe with IdentMen Typo DNIVIE and the elasticities}

\item{spending_var_df}{A spending_var datframe with SpendindgRef the sum of spending at the ref year, SpendingHor the target of spending at
the horizon}

\item{price_index_hor_df}{The price vector. Used for the IP Stone calculation}

\item{floor_at_z}{Boolean wether negative spendings are floored at zero}
}
\value{
A spending dataframe
}
\description{
This function will reprocess the budget in spending_econo until it is aligned with the column SpendingHor in spending_var
Uses the men_elast elasticities.
}
\examples{
ventilate_solde(spending_econo_df,  men_elast_df, spending_var_df, price_index_hor_df, floor_at_z = T)
}
