% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AlphaLauncher.R
\name{Step_6_FinalData}
\alias{Step_6_FinalData}
\title{Step_6_FinalData}
\usage{
Step_6_FinalData(MatisseData)
}
\arguments{
\item{MatisseData}{A MatisseData list}
}
\value{
A MatisseData list
}
\description{
The final data modification, given all former changes
}
\examples{
Step_6_FinalData(MatisseData)
}
